Removed old UI, replaced it with new
