--- UI Scaled Solution/master-script-form/README.md
@@ -1,1 +1,0 @@
-For instructions on how to install and deploy Angular UI, [refer to this file](https://github.com/k8-proxy/aws-jmeter-test-engine/blob/master/jmeter-icap/instructions/scaled-angular-ui-component-install-and-deploy.md).
