--- UI Scaled Solution/master-script-form/src/app/app.component.html
@@ -1,2 +1,0 @@
-<config-form></config-form>
-<tests-table></tests-table>
