--- UI Scaled Solution/master-script-form/src/app/config-form/config-form.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ConfigFormComponent } from './config-form.component';
-
-describe('ConfigFormComponent', () => {
-  let component: ConfigFormComponent;
-  let fixture: ComponentFixture<ConfigFormComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ ConfigFormComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ConfigFormComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
