--- UI Scaled Solution/master-script-form/src/styles.css
@@ -1,10 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-@import "~bootstrap/dist/css/bootstrap.css";
-
-body {
-    padding: 20px;
-}
-
-button {
-    margin-right: 20px;
-}
