+++ UI/master-script-form/src/app/app.component.html
@@ -1,1 +1,2 @@
 <config-form></config-form>
+<tests-table></tests-table>
