+++ UI/master-script-form/src/app/app.module.ts
@@ -9,11 +9,15 @@ import { AppComponent } from './app.component';
 import { ConfigFormComponent } from './config-form/config-form.component';
 import { HttpClientModule } from '@angular/common/http';
 import { ConfirmationPopoverModule } from 'angular-confirmation-popover';
+import { CookieService } from 'ngx-cookie-service';
+import { TestsTableComponent } from './tests-table/tests-table.component';
+import {MatTableModule} from '@angular/material/table';
 
 @NgModule({
   declarations: [
     AppComponent,
-    ConfigFormComponent
+    ConfigFormComponent,
+    TestsTableComponent,
   ],
   imports: [
     BrowserModule,
@@ -22,12 +26,13 @@ import { ConfirmationPopoverModule } from 'angular-confirmation-popover';
     ReactiveFormsModule,
     HttpClientModule,
     CommonModule,
+    MatTableModule,
     ConfirmationPopoverModule.forRoot({
       confirmButtonType: 'warning'
     })
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
-  providers: [FormBuilder],
+  providers: [FormBuilder, CookieService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
