+++ UI/master-script-form/src/app/common/Validators/ConfigFormValidators.ts
@@ -1,9 +1,10 @@
+import { AppSettings } from './../app settings/AppSettings';
 import { AbstractControl, ValidationErrors } from '@angular/forms';
 
 export class ConfigFormValidators {
 
     static hasNumberLimit(control: AbstractControl) : ValidationErrors | null {
-        if(control.value > 400) {
+        if(control.value > 4000) {
             return {exceedsNumberLimit: true};
         }
 
@@ -15,6 +16,13 @@ export class ConfigFormValidators {
             return {cannotContainSpaces: true}
         }
   
+        return null;
+    }
+
+    static cannotContainDuplicatePrefix(control: AbstractControl) : ValidationErrors | null {
+        if(control.value != null &&  AppSettings.testPrefixSet.has((control.value as string))){
+            return {cannotContainDuplicatePrefix: true}
+        }
         return null;
     }
 }
