+++ UI/master-script-form/src/app/common/app settings/AppSettings.ts
