+++ UI/master-script-form/src/app/common/services/shared.service.ts
