+++ UI/master-script-form/src/app/config-form/config-form.component.ts
@@ -1,3 +1,6 @@
+import { Subscription } from 'rxjs';
+import { AppSettings } from './../common/app settings/AppSettings';
+import { SharedService, FormDataPackage } from './../common/services/shared.service';
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder, FormControl, Validators } from '@angular/forms'
 import { Router } from '@angular/router';
@@ -5,7 +8,6 @@ import { HttpClient } from '@angular/common/http';
 import { Title } from '@angular/platform-browser';
 import { ConfigFormValidators } from '../common/Validators/ConfigFormValidators';
 import { animate, state, style, transition, trigger } from '@angular/animations';
-import { CookieService } from 'ngx-cookie-service';
 
 @Component({
   selector: 'config-form',
@@ -20,37 +22,42 @@ import { CookieService } from 'ngx-cookie-service';
     ])
   ]
 })
+
 export class ConfigFormComponent implements OnInit {
-  regions: string[] = ['eu-west-1', 'eu-east-1', 'us-west-1', 'eu-west-2'];
-  loadTypes: string[] = ['Direct', 'Proxy'];
-  urlChoices: string[] = ["ICAP Server Endpoint URL*", "Proxy IP Address*"];
+  testsStoppedSubscription: Subscription;
   configForm: FormGroup;
-  fileToUpload: File = null;
   submitted = false;
-  responseUrl = '';
   responseReceived = false;
   portDefault = '443';
   enableCheckboxes = true;
   enableIgnoreErrorCheckbox = true;
-  IcapOrProxy = this.urlChoices[0];
-  showStoppedAlert = false;
+  IcapOrProxy = AppSettings.urlChoices[0];
+  showErrorAlert = false;
   hideSubmitMessages = false;
-  public popoverTitle: string = "Please Confirm";
-  public popoverMessage: string = "Are you sure you wish to stop all load?";
-  public confirmClicked: boolean = false;
-  public cancelClicked: boolean = false;
+  GenerateLoadButtonText = "Generate Load";
 
-  constructor(private fb: FormBuilder, private readonly http: HttpClient, private router: Router, private titleService: Title, public cookieService: CookieService) { }
+  constructor(private fb: FormBuilder, private readonly http: HttpClient, private router: Router, private titleService: Title, private sharedService: SharedService) {
+    this.testsStoppedSubscription = this.sharedService.getStopSingleEvent().subscribe((prefix) => this.onTestStopped(prefix));
+  }
 
   ngOnInit(): void {
     this.initializeForm();
     this.setTitle("ICAP Performance Test");
-    console.log(this.cookieService.getAll());
     this.configForm.valueChanges.subscribe((data) => {
       this.hideSubmitMessages = true;
     });
-    setInterval(() => { this.getCookies(); }, 1000); //used to refresh list and remove expired tests.
-    
+    this.setIcapOrProxyValidation();
+  }
+
+  setIcapOrProxyValidation() {
+    this.configForm.get('load_type').valueChanges.subscribe(loadType => {
+      if (loadType == "Proxy") {
+        this.icap_endpoint_url.setValidators([Validators.required, ConfigFormValidators.cannotContainSpaces, Validators.pattern(/^(([1-9]?\d|1\d\d|2[0-5][0-5]|2[0-4]\d)\.){3}([1-9]?\d|1\d\d|2[0-5][0-5]|2[0-4]\d)$/)]);
+      } else {
+        this.icap_endpoint_url.setValidators([Validators.required, ConfigFormValidators.cannotContainSpaces]);
+      }
+      this.configForm.get('icap_endpoint_url').updateValueAndValidity();
+    })
   }
 
   setTitle(newTitle: string) {
@@ -62,9 +69,9 @@ export class ConfigFormComponent implements OnInit {
       total_users: new FormControl('', [Validators.pattern(/^(?=.*\d)[\d ]+$/), ConfigFormValidators.cannotContainSpaces, ConfigFormValidators.hasNumberLimit]),
       duration: new FormControl('', [Validators.pattern(/^(?=.*\d)[\d ]+$/), ConfigFormValidators.cannotContainSpaces]),
       ramp_up_time: new FormControl('', [Validators.pattern(/^(?=.*\d)[\d ]+$/), ConfigFormValidators.cannotContainSpaces]),
-      load_type: this.loadTypes[0],
+      load_type: AppSettings.loadTypes[0],
       icap_endpoint_url: new FormControl('', [Validators.required, ConfigFormValidators.cannotContainSpaces]),
-      prefix: new FormControl('', [ConfigFormValidators.cannotContainSpaces]),
+      prefix: new FormControl('', [ConfigFormValidators.cannotContainSpaces, ConfigFormValidators.cannotContainDuplicatePrefix, Validators.required]),
       enable_tls: true,
       tls_ignore_error: true,
       port: new FormControl('', [Validators.pattern(/^(?=.*\d)[\d ]+$/), ConfigFormValidators.cannotContainSpaces]),
@@ -73,15 +80,14 @@ export class ConfigFormComponent implements OnInit {
 
   onLoadTypeChange() {
     //if direct, else proxy
-    if (this.configForm.get('load_type').value == this.loadTypes[0]) {
+    if (this.configForm.get('load_type').value == AppSettings.loadTypes[0]) {
       this.enableCheckboxes = true;
-      this.IcapOrProxy = this.urlChoices[0];
-    } else if (this.configForm.get('load_type').value == this.loadTypes[1]) {
+      this.IcapOrProxy = AppSettings.urlChoices[0];
+    } else if (this.configForm.get('load_type').value == AppSettings.loadTypes[1]) {
       this.enableCheckboxes = false;
-      this.IcapOrProxy = this.urlChoices[1];
+      this.IcapOrProxy = AppSettings.urlChoices[1];
     }
   }
-
 
   onTlsChange() {
     if (this.configForm.get('enable_tls').value == true) {
@@ -112,45 +118,34 @@ export class ConfigFormComponent implements OnInit {
   get prefix() {
     return this.configForm.get('prefix');
   }
-
   get isValid() {
     return this.configForm.valid;
   }
-
   get formSubmitted() {
     return this.submitted;
   }
-
   get gotResponse() {
     return this.responseReceived;
   }
-
-  get getUrl() {
-    return this.responseUrl;
-  }
-
   get animState() {
-    return this.showStoppedAlert ? 'show' : 'hide';
+    return this.showErrorAlert ? 'show' : 'hide';
+  }
+  get cookiesExist(): boolean {
+    return AppSettings.cookiesExist;
+  }
+  get loadTypes() {
+    return AppSettings.loadTypes;
   }
 
-  onFileChange(files: FileList) {
-    this.fileToUpload = files.item(0);
-  }
-
-  processResponse(response: object) {
-    this.responseUrl = response.toString();
+  processResponse(response: object, formData: FormData) {
+    let formAsString = formData.get('form');
     this.responseReceived = true;
-    this.storeTestAsCookie(this.responseUrl);
-  }
 
-  
-
-  storeTestAsCookie(dashboardUrl) {
-    let currentTime = new Date();
-    let expireTime = new Date(currentTime.getTime() + this.duration.value * 1000);
-    let testTitle = this.IcapOrProxy === this.urlChoices[0] ? "ICAP Live Performance Dashboard" : "Proxy Site Live Performance Dashboard";
-    let key = this.prefix.value === null ? testTitle : this.prefix.value + " " + testTitle;
-    this.cookieService.set(key, dashboardUrl, expireTime);
+    //pack up form data and response URL, fire form submitted event and send to subscribers
+    const dataPackage: FormDataPackage = { formAsJsonString: formAsString.toString(), grafanaUrlResponse: response['url'], stackName: response['stack_name'] }
+    this.sharedService.sendSubmitEvent(dataPackage);
+    this.unlockForm();
+    this.submitted = false;
   }
 
   resetForm() {
@@ -161,7 +156,7 @@ export class ConfigFormComponent implements OnInit {
   }
 
   postFormToServer(formData: FormData) {
-    this.http.post('http://127.0.0.1:5000/', formData).subscribe(response => this.processResponse(response));
+    this.http.post('http://127.0.0.1:5000/', formData).subscribe(response => this.processResponse(response, formData), (err) => { this.onError(err) });
   }
 
   postStopRequestToServer(formData: FormData) {
@@ -172,64 +167,72 @@ export class ConfigFormComponent implements OnInit {
     this.setFormDefaults();
     this.hideSubmitMessages = false;
     if (this.configForm.valid) {
+      AppSettings.addingPrefix = true;
+      AppSettings.testPrefixSet.add(this.prefix.value);
       //append the necessary data to formData and send to Flask server
       const formData = new FormData();
       formData.append("button", "generate_load");
-      if (this.fileToUpload) {
-        formData.append('file', this.fileToUpload, this.fileToUpload.name);
-      }
       formData.append('form', JSON.stringify(this.configForm.getRawValue()));
       this.postFormToServer(formData);
       this.submitted = true;
+      this.lockForm();
     }
+  }
+
+  lockForm() {
+    this.GenerateLoadButtonText = "Generating Load..."
+    this.configForm.disable();
+  }
+
+  unlockForm() {
+    this.GenerateLoadButtonText = "Generate Load"
+    this.configForm.enable();
+    this.prefix.reset();
   }
 
   setFormDefaults() {
     //if user enters less that 1 total_users, default to 1. Otherwise if no input, default to 25.
-    if(this.total_users.value === '') {
+    if (this.total_users.value === '') {
       this.total_users.setValue('25');
     } else if (this.total_users.value < 1) {
       this.total_users.setValue('1');
-    } 
+    }
 
     //if user enters no ramp up time, default is 300.
-    if(this.ramp_up_time.value === '') {
+    if (this.ramp_up_time.value === '') {
       this.ramp_up_time.setValue('300');
     }
 
     //if user enters no duration, default is 900. If they enter a less than 60 second duration, default to 60.
-    if(this.duration.value === '') {
+    if (this.duration.value === '') {
       this.duration.setValue('900');
     }
     else if (this.duration.value < 60) {
       this.duration.setValue('60');
     }
-
-    if(this.prefix.value === '') {
-      this.prefix.setValue('demo');
-    }
   }
 
-  onStopTests() {
-    const formData = new FormData();
-    formData.append("button", "stop_tests");
-    this.postStopRequestToServer(formData);
-    this.cookieService.deleteAll();
-    this.toggleTerminationAlert();
+  onError(error) {
+    console.log(error);
+    this.toggleErrorMessage();
     this.submitted = false;
     this.responseReceived = false;
-    setTimeout(() => this.toggleTerminationAlert(), 3000);
+    setTimeout(() => this.toggleErrorMessage(), 3000);
+    this.unlockForm();
+    AppSettings.addingPrefix = false;
   }
 
-  toggleTerminationAlert() {
-    this.showStoppedAlert = !this.showStoppedAlert;
+  toggleErrorMessage() {
+    this.showErrorAlert = !this.showErrorAlert;
   }
 
-  cookiesExist(): boolean {
-    return !(Object.keys(this.cookieService.getAll()).length === 0 && this.cookieService.getAll().constructor === Object);
-  }
-
-  getCookies() {
-    return this.cookieService.getAll();
+  //used to revalidate prefix if a test is stopped. So in instances where a prefix is invalid due to an existing test, it being deleted will make that prefix valid again.
+  onTestStopped(prefix: string) {
+    if (this.prefix.value === prefix) {
+      this.prefix.markAsPristine();
+      this.prefix.markAsUntouched();
+      this.prefix.updateValueAndValidity();
+      this.configForm.updateValueAndValidity();
+    }
   }
 }
